feat/lazy-loading: add lazy-loading for history page and fix user store
