+++ src/app/[locale]/layout.tsx
@@ -25,9 +25,7 @@ export default async function RootLayout({
   }
 
   const supabase = await createServerSupabase();
-
   const { data } = await supabase.auth.getUser();
-  const userName = data?.user?.user_metadata.username;
 
   // Enable static rendering
   setRequestLocale(locale);
@@ -37,7 +35,7 @@ export default async function RootLayout({
     <NextIntlClientProvider messages={messages}>
       <AppRouterCacheProvider>
         <ThemeProvider theme={theme}>
-          <Header initialUserName={userName} />
+          <Header initialUser={data?.user} />
           {children}
           <Footer />
           <SnackbarContainer />
