+++ src/store/userAuth/userAuth-store.tsx
@@ -1,11 +1,12 @@
+import { User } from '@supabase/supabase-js';
 import { create } from 'zustand';
 
 type userAuthStore = {
-  userName: string | null;
-  setUserName: (name: string | null) => void;
+  userData: User | null;
+  setUser: (user: User | null) => void;
 };
 
 export const userAuthStore = create<userAuthStore>((set) => ({
-  userName: null,
-  setUserName: (name) => set({ userName: name }),
+  userData: null,
+  setUser: (user: User | null) => set({ userData: user }),
 }));
